Created add_data() method
I finished the add_data() method which allows adding new data to the blockchain.
It creates a new block with the user’s data, runs proof of work to mine it,
and then adds it to the chain. This is the main way to add blocks based on user input.