Completed is_chain_valid() method
I added the is_chain_valid() function to make sure the blockchain hasn’t been tampered with. It checks that each block’s hash is still correct and that every block is properly connected to the previous one. If something has been changed, it will print an error and return False.